+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/Main.scala
@@ -9,12 +9,13 @@ import io.iohk.atala.pollux.sql.repository.{Migrations => PolluxMigrations}
 import io.iohk.atala.connect.sql.repository.{Migrations => ConnectMigrations}
 
 object Main extends ZIOAppDefault {
-  def agentLayer(peer: PeerDID): ZLayer[Any, Nothing, AgentServiceAny] = ZLayer.succeed(
-    io.iohk.atala.mercury.AgentServiceAny(
-      new DIDComm(UniversalDidResolver, peer.getSecretResolverInMemory),
-      peer.did
+  def agentLayer(peer: PeerDID): ZLayer[Any, Nothing, AgentServiceAny] =
+    ZLayer.succeed(
+      io.iohk.atala.mercury.AgentServiceAny(
+        new DIDComm(UniversalDidResolver, peer.getSecretResolverInMemory),
+        peer.did
+      )
     )
-  )
 
   override def run: ZIO[Any, Throwable, Unit] =
     for {
@@ -42,6 +43,12 @@ object Main extends ZIOAppDefault {
       }
       _ <- ZIO.logInfo(s"REST Service port => $restServicePort")
 
+      didCommServiceUrl <- System.env("DIDCOMM_SERVICE_URL").map {
+        case Some(s) => s
+        case _       => "http://localhost"
+      }
+      _ <- ZIO.logInfo(s"DIDComm Service URL => $didCommServiceUrl")
+
       didCommServicePort <- System.env("DIDCOMM_SERVICE_PORT").map {
         case Some(s) if s.toIntOption.isDefined => s.toInt
         case _                                  => 8090
@@ -57,7 +64,7 @@ object Main extends ZIOAppDefault {
         .provide(RepoModule.connectDbConfigLayer >>> ConnectMigrations.layer)
 
       agentDID <- for {
-        peer <- ZIO.succeed(PeerDID.makePeerDid(serviceEndpoint = Some(s"http://localhost:$didCommServicePort")))
+        peer <- ZIO.succeed(PeerDID.makePeerDid(serviceEndpoint = Some(s"$didCommServiceUrl:$didCommServicePort")))
         _ <- ZIO.logInfo(s"New DID: ${peer.did}") *>
           ZIO.logInfo(s"JWK for KeyAgreement: ${peer.jwkForKeyAgreement.toJSONString}") *>
           ZIO.logInfo(s"JWK for KeyAuthentication: ${peer.jwkForKeyAuthentication.toJSONString}")