+++ prism-agent/service/server/src/main/scala/io/iohk/atala/api/http/FailureResponse.scala
@@ -2,6 +2,7 @@ package io.iohk.atala.api.http
 
 import sttp.model.StatusCode
 import sttp.tapir.EndpointOutput.OneOf
+import sttp.tapir.Schema
 import sttp.tapir.generic.auto.*
 import sttp.tapir.json.zio.jsonBody
 import zio.json.{DeriveJsonDecoder, DeriveJsonEncoder}
@@ -13,6 +14,7 @@ case class NotFoundResponse(msg: String) extends FailureResponse
 object NotFoundResponse {
   given encoder: zio.json.JsonEncoder[NotFoundResponse] = DeriveJsonEncoder.gen[NotFoundResponse]
   given decoder: zio.json.JsonDecoder[NotFoundResponse] = DeriveJsonDecoder.gen[NotFoundResponse]
+  given schema: Schema[NotFoundResponse] = Schema.derived
 }
 
 case class BadRequest(msg: String, errors: List[String] = List.empty) extends FailureResponse
@@ -20,6 +22,7 @@ case class BadRequest(msg: String, errors: List[String] = List.empty) extends Fa
 object BadRequest {
   given encoder: zio.json.JsonEncoder[BadRequest] = DeriveJsonEncoder.gen[BadRequest]
   given decoder: zio.json.JsonDecoder[BadRequest] = DeriveJsonDecoder.gen[BadRequest]
+  given schema: Schema[BadRequest] = Schema.derived
 }
 
 case class InternalServerError(msg: String) extends FailureResponse
@@ -27,6 +30,7 @@ case class InternalServerError(msg: String) extends FailureResponse
 object InternalServerError {
   given encoder: zio.json.JsonEncoder[InternalServerError] = DeriveJsonEncoder.gen[InternalServerError]
   given decoder: zio.json.JsonDecoder[InternalServerError] = DeriveJsonDecoder.gen[InternalServerError]
+  given schema: Schema[InternalServerError] = Schema.derived
 }
 
 //An RFC-7807 compliant data structure for reporting errors to the client
@@ -36,4 +40,5 @@ case class ErrorResponse(`type`: String, title: String, status: Int, instance: S
 object ErrorResponse {
   given encoder: zio.json.JsonEncoder[ErrorResponse] = DeriveJsonEncoder.gen[ErrorResponse]
   given decoder: zio.json.JsonDecoder[ErrorResponse] = DeriveJsonDecoder.gen[ErrorResponse]
+  given schema: Schema[ErrorResponse] = Schema.derived
 }